Error in the format of the ossec.conf files and command entries only added on the manager side.
Error in the format of the ossec.conf files and command entries only added on the manager side.

We use the Nokogiri gem to obtain the desired format after using Gyoku.

We substitute the value of "all" for the "server" in the attributes of the ossec.conf files so that it only affects the manager.